HR: New test QR code added - 1/1 vaccination certificate for 2-dosed vaccine (#78)
* Add files via upload

* Delete HR/1.3.0 directory

* PNG files uploaded, jpeg deleted

* New test QR code added - 1/1 vaccination certificate for double dose vaccine

we are submitting one additional code for testing - a 1/1 vaccine certificate for a 2-dosed vaccine (Moderna in this case). We want to get a confrimation trough the round 3 testing that this certificate can be validated trough out MS.
it represents a full vaccination after a recovery (pcr in 180 days) and 1/2 doses of received vaccine.